Enum r::object::Vector

pub enum Vector {
    Double(Rep<Double>),
    Integer(Rep<Integer>),
    Logical(Rep<Logical>),
    Character(Rep<Character>),
}

Variants

Double(Rep<Double>)

Integer(Rep<Integer>)

Logical(Rep<Logical>)

Character(Rep<Character>)

Implementations

impl Vector

pub fn get(&self, index: usize) -> Option<Vector>

pub fn lazy_copy(&self) -> Self

Create a lazy copy of the vector. When mutating vectors, the internal data only needs to be copied when there is more than one such lazy copy.

pub fn try_get(&self, index: Obj) -> EvalResult

pub fn subset(&self, subset: Subset) -> Self

pub fn assign(&mut self, other: Obj) -> EvalResult

pub fn materialize(self) -> Self

pub fn vec_coerce<T, U>(v: &[OptionNA<T>]) -> Vec<OptionNA<U>>
where T: CoercibleInto<U> + Clone,

pub fn vec_parse<U>(v: &[OptionNA<String>]) -> (bool, Vec<OptionNA<U>>)
where U: FromStr,

pub fn as_integer(self) -> Vector

pub fn as_double(self) -> Vector

pub fn as_logical(self) -> Vector

pub fn as_character(self) -> Vector

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool