Function r::object::iterators::map_common_numeric

pub fn map_common_numeric<I, LItem, RItem, DLItem, DRItem, LNum, RNum, Output>(
    i: I,
) -> impl Iterator<Item = (Output, Output)>
where
    I: IntoIterator<Item = (LItem, RItem)>,
    LItem: MinimallyNumeric<As = LNum> + Deref<Target = DLItem>,
    RItem: MinimallyNumeric<As = RNum> + Deref<Target = DRItem>,
    DLItem: CoercibleInto<LNum> + Clone,
    DRItem: CoercibleInto<RNum> + Clone,
    (LNum, RNum): CommonNum<Common = Output>,

Map an iterator of pairs into a pair of common numeric types

Accept an iterator of pairs of numeric (or numeric-coercible) values and returns an iterator of pairs of numerics coerced to the least greater common type, assuming a hierarchy of representations.

Arguments

i - An iterator over pairs of numeric (or numeric-coercible) values