Enum r::object::rep::Rep

pub enum Rep<T: Clone> {
    Subset(CowObj<Vec<T>>, Subsets, Option<Naming>),
}

Vector

Variants

Subset(CowObj<Vec<T>>, Subsets, Option<Naming>)

Implementations

impl<T: Clone + Default + ViewMut> Rep<T>

pub fn try_get_inner(&self, subset: Subset) -> Result<T, Signal>

Get a cloned version of the inner value. This is used for accessing inner values like list(1)[[1]].

pub fn try_get_inner_mut(&self, subset: Subset) -> Result<T, Signal>

Retrieve the internal data as a mutable view. This is important for lists for things like l$a[1:2] = c(10, 11)

impl<T: Clone + Default> Rep<T>

pub fn new() -> Self

Create an empty vector

The primary use case for this function is to support testing, and there are few expected use cases outside. It is used for creating a vector of an explicit atomic type, likely to be tested with SameType::is_same_type_as.

use r::utils::*;
use r::object::Vector;
use r::object::OptionNA;

let result = Vector::from(vec![1, 2, 3]);
let expect = Vector::from(Vec::<OptionNA<i32>>::new());

assert!(result.is_same_type_as(&expect))

pub fn is_named(&self) -> bool

Whether the vector representation has names.

pub fn names(&self) -> Option<CowObj<Vec<Character>>>

Return the names of the vector if there are any.

pub fn set_subset(&mut self, subset: Subset, value: T) -> Result<T, Signal>

Change a value at the location given by subset to the provided value. If the subset does not have length 1, an error is returned.

pub fn values_ref(&self) -> IntoIterableRefValues<T>

Get an IntoIterableValues<T> which in turn can be converted into an iterator over references to the values (&T).

Directly getting an iterator is not possible due to lifetime issues.

pub fn names_ref(&self) -> Option<IntoIterableRefNames>

Get an Option<IntoIterableRefNames> which in turn can be converted into an iterator over references to the names (&String). None is returned when no names exist.

Directly getting an iterator is not possible due to lifetime issues.

pub fn pairs_ref(&self) -> IntoIterableRefPairs<T>

Get an IntoIterablePairs<T> which in turn can be converted into an iterator over pairs of references ((&String, &T)).

Directly getting an iterator is not possible due to lifetime issues.

pub fn iter_pairs(&self) -> IterablePairs<T>

Iterate over (owned) pairs of names and values ((String, T)).

pub fn iter_values(&self) -> IterableValues<T>

Iterate over the (owned) values of the vector.

pub fn iter_names(&self) -> Option<IterableValues<Character>>

Iterate over the names of the vector (if they exist).

pub fn push_value(&mut self, value: T)

pub fn push_named(&mut self, name: OptionNA<String>, value: T)

Push a named value with a given name onto the Rep<T>.

pub fn iter_subset_indices_exact(&self) -> ExactIterSubsetIndices

pub fn iter_subset_indices(&self) -> Box<dyn Iterator<Item = Option<usize>>>

pub fn reindex(&mut self)

Reindex the mapping from names to indices.

pub fn with_capacity(capacity: usize, names: bool) -> Self

Constructs a new, empty Rep<T> with at least the specified capacity. Names are only include if names is true.

pub fn dedup_last(self) -> Self

pub fn set_names(&self, names: CowObj<Vec<Character>>) -> Self

pub fn inner(&self) -> CowObj<Vec<T>>

Access a lazy copy of the internal vector data

pub fn with_inner_mut<F, R>(&self, f: F) -> R
where F: FnOnce(&mut Vec<T>) -> R,

Get mutable access to the internal vector through the passed closure.

pub fn subset(&self, subset: Subset) -> Self

Subsetting a Vector

Introduce a new subset into the aggregate list of subset indices.

pub fn len(&self) -> usize

The length of the vector.

pub fn is_empty(&self) -> bool

Whether the vector has length 0.

pub fn get(&self, index: usize) -> Option<Rep<T>>
where T: Clone,

Get a single element from a vector

Access a single element without materializing a new vector

pub fn assign<R>(&mut self, value: Rep<R>) -> Result<Self, Signal>
where T: Clone + Default + From<R>, R: Default + Clone,

Assignment to Subset Indices

Assignment to a vector from another. The aggregate subsetted indices are iterated over while performing the assignment.

pub fn as_scalar(&self) -> Option<T>

Return the only value if the vector has length 1.

pub fn materialize(&self) -> Self
where T: Clone,

Materialize a Vector

Apply subsets and clone values into a new vector.

pub fn is_double(&self) -> bool
where T: AtomicMode,

Test the mode of the internal vector type

Internally, this is defined by the crate::object::coercion::AtomicMode implementation of the vector’s element type.

pub fn is_logical(&self) -> bool
where T: AtomicMode,

See Self::is_double for more information

pub fn is_integer(&self) -> bool
where T: AtomicMode,

See Self::is_double for more information

pub fn is_character(&self) -> bool
where T: AtomicMode,

See Self::is_double for more information

pub fn as_mode<Mode>(&self) -> Rep<Mode>
where T: CoercibleInto<Mode>, Mode: Clone,

Convert a Vector into a vector of a specific class of internal type

The internal type only needs to satisfy crate::object::coercion::CoercibleInto for the Mode, and for the Mode type to implement crate::object::coercion::AtomicMode. Generally, this is used more directly via Self::as_logical, Self::as_integer, Self::as_double and Self::as_character, which predefine the output type of the mode.

use r::object::Vector;
use r::object::OptionNA;

let x = Vector::from(vec![false, true, true, false]);
let n = x.as_double();

assert_eq!(n, Vector::from(vec![
   OptionNA::Some(0_f64),
   OptionNA::Some(1_f64),
   OptionNA::Some(1_f64),
   OptionNA::Some(0_f64)
]))

pub fn as_logical(&self) -> Rep<Logical>
where T: CoercibleInto<Logical>,

See Self::as_mode for more information

pub fn as_integer(&self) -> Rep<Integer>
where T: CoercibleInto<Integer>,

See Self::as_mode for more information

pub fn as_double(&self) -> Rep<Double>
where T: CoercibleInto<Double>,

See Self::as_mode for more information

pub fn as_character(&self) -> Rep<Character>
where T: CoercibleInto<Character>,

See Self::as_mode for more information

pub fn get_inner(&self, index: usize) -> Option<T>