Struct r::object::rep::Rep

pub struct Rep<T: Clone>(pub RefCell<RepType<T>>);

Vector Representation

The ref-cell is used so vectors can change there internal representation, e.g. by materializing.

Tuple Fields
0: RefCell<RepType<T>>
Implementations

impl<T> Rep<T>
where T: AtomicMode + Clone + Default,

pub fn materialize(&self) -> Self

pub fn new() -> Self

Create an empty vector

The primary use case for this function is to support testing, and there are few expected use cases outside. It is used for creating a vector of an explicit atomic type, likely to be tested with SameType::is_same_type_as.

use r::utils::*;
use r::object::Vector;
use r::object::OptionNA;

let result = Vector::from(vec![1, 2, 3]);
let expect = Vector::from(Vec::<OptionNA<i32>>::new());

assert!(result.is_same_type_as(&expect))

pub fn inner(&self) -> CowObj<Vec<T>>

pub fn len(&self) -> usize

pub fn subset(&self, subset: Subset) -> Self

Subsetting a Vector

Introduce a new subset into the aggregate list of subset indices.

pub fn is_empty(&self) -> bool

pub fn get(&self, index: usize) -> Option<Self>

pub fn assign(&mut self, value: Self) -> Self

pub fn is_double(&self) -> bool

Test the mode of the internal vector type

Internally, this is defined by the crate::object::coercion::AtomicMode implementation of the vector’s element type.

pub fn is_logical(&self) -> bool

See Self::is_double for more information

pub fn is_integer(&self) -> bool

See Self::is_double for more information

pub fn is_character(&self) -> bool

See Self::is_double for more information

pub fn as_mode<Mode>(&self) -> Rep<Mode>
where T: CoercibleInto<Mode>, Mode: Clone,

Convert a Vector into a vector of a specific class of internal type

The internal type only needs to satisfy crate::object::coercion::CoercibleInto for the Mode, and for the Mode type to implement crate::object::coercion::AtomicMode. Generally, this is used more directly via Self::as_logical, Self::as_integer, Self::as_double and Self::as_character, which predefine the output type of the mode.

use r::object::Vector;
use r::object::OptionNA;

let x = Vector::from(vec![false, true, true, false]);
let n = x.as_double();

assert_eq!(n, Vector::from(vec![
   OptionNA::Some(0_f64),
   OptionNA::Some(1_f64),
   OptionNA::Some(1_f64),
   OptionNA::Some(0_f64)
]))

pub fn as_logical(&self) -> Rep<Logical>
where T: CoercibleInto<Logical>,

See Self::as_mode for more information

pub fn as_integer(&self) -> Rep<Integer>
where T: CoercibleInto<Integer>,

See Self::as_mode for more information

pub fn as_double(&self) -> Rep<Double>
where T: CoercibleInto<Double>,

See Self::as_mode for more information

pub fn as_character(&self) -> Rep<Character>
where T: CoercibleInto<Character>,

See Self::as_mode for more information

pub fn vectorized_partial_cmp<R, C>( self, other: Rep<R>, ) -> Vec<Option<Ordering>>
where T: AtomicMode + Default + Clone + CoercibleInto<C>, R: AtomicMode + Default + Clone + CoercibleInto<C>, (T, R): CommonCmp<Common = C>, C: PartialOrd,

Apply over the vector contents to produce a vector of std::cmp::Ordering

This function is used primarily in support of the implementation of vectorized comparison operators and likely does not need to be used outside of that context.

See crate::object::vector::VecPartialCmp for vectorized comparison operator implementations.